A helper function to help parse API calls from the census

The function verifies that API call was sucessful. If the call was not sucessful, this function passes the message received from the US Census API for further troublshooting,

check_census_api_call(call)

Arguments

call

a returned call from the US Census API

Value

a string vector with the message from the US Census API

Examples

# \donttest{
library(tidyqwi)
library(httr)
# A single call to the API without an API Key
url <- "api.census.gov/data/timeseries/qwi/sa?get=Emp&for=county:198&key=NOKEY"
single_call <- httr::GET(url)

# Because a non valid API key was specified an message will be returned

check_census_api_call(single_call)
#> Error in check_census_api_call(single_call): 
#> The following is a message from the US Census API:
#> 
#> <html style="font-size: 14px;">
#>     <head>
#>         <title>Invalid Key</title>
#>         <link rel="stylesheet" type="text/css" href="assets/styles.css">
#>         <script type="text/javascript" src="assets/jquery-1.4.4.min.js"></script>
#>         <script type="text/javascript">
#>             $(document).ready(function() {
#>                 $(".menu-activator").click(function() {
#>                     $(".gov-menu").toggle()
#>                     $(".chevron").toggleClass('rotate')
#>                 })                
#>             })
#>         </script>
#>     </head>
#>     <body>
#>         <header>
#>             <div data-v-8adc6fed="" data-v-2f72e816="" class="gov-banner noprint">
#>                 <div data-v-8adc6fed="" class="aqua-layout horizontal align-start justify-start banner-header">
#>                     <img data-v-8adc6fed="" src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAALCAMAAABBPP0LAAAAG1BMVEUdM7EeNLIeM7HgQCDaPh/bPh/bPx/////bPyBEby41AAAAUElEQVQI123MNw4CABDEwD3jC/9/MQ1BQrgeOSkIqYe2o2FZtthXgQLgbHVMZdlsfUQFQnHtjP1+8BUhBDKOqtmfot6ojqPzR7TjdU+f6vkED+IDPhTBcMAAAAAASUVORK5CYII=" alt="U.S flag" class="usflag" role="img" aria-hidden="true">
#>                     <span data-v-8adc6fed="">
#>                         <div data-v-8adc6fed="" class="aqua-layout horizontal" tabindex="0">
#>                             <div data-v-8adc6fed="" tabindex="0" aria-label="An official website of the United States government"> 
#>                                 An official website of the United States government 
#>                             </div>
#>                             <div data-v-8adc6fed="" class="menu-activator" aria-label="Here's how you know. Press enter for more information." tabindex="0">
#>                                 <div data-v-8adc6fed="" class="aqua-layout horizontal justify-start"> Here&rsquo;s how you know <div data-v-8adc6fed="" class="chevron"><img  data-v-8adc6fed="" class="aqua-icon" src="assets/images/chevron-down.svg" style="width: 0.75rem; height: 0.75rem; max-height: 0.75rem; transition: all 500ms ease 0s; filter: invert(30%) sepia(8%) saturate(3837%) hue-rotate(171deg) brightness(99%) contrast(89%);"></div>
#>                                 </div>
#>                             </div>
#>                         </div>
#>                     </span>
#>                 </div>
#>                 <div data-v-8adc6fed="" class="aqua-layout horizontal justify-start gov-menu" style="display: none;"><div data-v-8adc6fed="" class="aqua-layout horizontal justify-start menu-paragraph"><img data-v-8adc6fed="" class="menu-svg" src="assets/images/icon-gov-building.svg" role="img" alt="government building image" aria-hidden="true"><div data-v-8adc6fed="" tabindex="0" aria-label="Official websites use .gov. A .gov website belongs to an official government
#>                 organization in the United States."><strong data-v-8adc6fed=""> Official websites use .gov </strong>
#>                 <br data-v-8adc6fed=""> A <strong data-v-8adc6fed="">.gov</strong> website belongs to an official government organization in the United States. </div>
#>             </div>
#>             <div data-v-8adc6fed="" class="aqua-layout horizontal justify-start menu-paragraph"><img data-v-8adc6fed="" class="menu-svg" src="assets/images/icon-lock.svg" role="img" alt="lock image" aria-hidden="true">
#>                 <div data-v-8adc6fed="" tabindex="0" aria-label="Secure .gov websites use HTTPS. A lock or https:// means you’ve safely connected to the
#>                 .gov website. Share sensitive information only on official, secure
#>                 websites."><strong data-v-8adc6fed=""> Secure .gov websites use HTTPS </strong>
#>                 <br data-v-8adc6fed=""> A <strong data-v-8adc6fed="">lock</strong> ( <img data-v-8adc6fed="" class="menu-lock" src="assets/images/lock.svg" role="img" alt="lock image" aria-hidden="true"> ) or <strong data-v-8adc6fed="">https://</strong> means you&lsquo;ve safely connected to the .gov website. Share sensitive information only on official, secure websites. </div>
#>             </div>
#>         </div>
#>     </div>
#>     <div data-v-beed8774="" data-v-2f72e816="" role="banner" class="AppHeader noprint" aqua-relative="">
#>         <div data-v-beed8774="" class="aqua-layout horizontal align-center justify-center"><!---->
#>             <div data-v-beed8774="" class="aqua-flex" style="flex: 0 0 auto;">
#>                 <div data-v-beed8774="" class="aqua-layout vertical">
#>                     <div data-v-beed8774="" class="logoArea">
#>                         <a data-v-fb9b02c5="" data-v-beed8774="" href="/" class="" id="census-home-link">
#>                             <div data-v-fb9b02c5="" style="display: none;"></div>
#>                             <img data-v-beed8774="" class="header-logo" src="assets/images/census-logo-gray.svg" alt="United States Census Bureau - Census Data Homepage">
#>                         </a>
#>                     </div>
#>                 </div>
#>             </div><!----><!---->
#>             <div data-v-beed8774="" class="aqua-layout vertical"></div>
#>         </div>
#>     </div>
#> </div>
#> </header>
#> <div class="flex-container content">
#>     <div class="content-wrap center">
#>         <h1>Invalid Key</h1>
#>         <p>
#>             A valid <em>key</em> must be included with each data API request.
#>             You included a key with this request, however, it is not valid.
#>             Please check your key and try again.
#>         </p>
#>         <p>
#>             If you do not have a key you may sign up for one <a href="key_signup.html">here</a>.
#>         </p>
#>     </div>      
#> </div>
#> 
#> <div data-v-56f1d3f7="" class="aqua-flex" style="flex: 0 0 auto;">
#>     <footer data-v-db923c0a="" data-v-56f1d3f7="" class="PageFooter">
#>         <div data-v-db923c0a="" class="aqua-layout align-center justify-center links aqua-text-caption">
#>             <a data-v-db923c0a="" target="_blank" href="https://www.census.gov/about/policies/section-508.html">Accessibility</a>
#>             <span data-v-db923c0a="" aria-hidden="true">&nbsp;|&nbsp;</span>
#>             <a data-v-db923c0a="" target="_blank" href="https://www.census.gov/quality/">Information Quality</a>
#>             <span data-v-db923c0a="" aria-hidden="true">&nbsp;|&nbsp;</span>
#>             <a data-v-db923c0a="" target="_blank" href="https://www.census.gov/foia/">FOIA</a>
#>             <span data-v-db923c0a="" aria-hidden="true">&nbsp;|&nbsp;</span>
#>             <a data-v-db923c0a="" target="_blank" href="https://www.census.gov/privacy/">Data Protection and Privacy Policy</a>
#>             <span data-v-db923c0a="" aria-hidden="true">&nbsp;|&nbsp;</span>
#>             <a data-v-db923c0a="" target="_blank" href="https://www.commerce.gov/">U.S. Department of Commerce</a>
#>             <span data-v-db923c0a="" aria-hidden="true">&nbsp;|&nbsp;</span>
#>             <a data-v-db923c0a="" target="_blank" href="https://www2.census.gov/data/api-documentation/data-census-gov-release-notes.pdf?#">Release Notes</a>
#>         </div>
#>     </footer>
#> </div>
#> 
#> 
#>     </body>
#> </html>

# }